requirements.txt
# Core dependencies
python-dotenv>=1.0.0
python-telegram-bot[ext]>=20.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# No additional paid dependencies required
# All functionality uses free APIs and services
